Enabled Profile view option

+++ app/profile/page.js
@@ -22,6 +22,7 @@ const MyProfile = () => {
       }
   
       if(session?.user.id) fetchPosts()
+    // }, [session?.user.id])
     }, [])
 
     const handleEdit = (post) => {
@@ -30,17 +31,18 @@ const MyProfile = () => {
     
     const handleDelete = async (post) => {
       const hasConfirmed = confirm("Are you sure you want to delete this prompt?")
-
-      try {
-        await fetch(`/api/prompt/${post._id.toString()}`, {
-          method: 'DELETE'
-        })
-
-        const filteredPosts = posts.filter((p) => p._id !== post._id)
-
-        setPosts(filteredPosts)
-      } catch (error) {
-        console.log(error)
+      if(hasConfirmed){
+        try {
+          await fetch(`/api/prompt/${post._id.toString()}`, {
+            method: 'DELETE'
+          })
+  
+          const filteredPosts = posts.filter((p) => p._id !== post._id)
+  
+          setPosts(filteredPosts)
+        } catch (error) {
+          console.log(error)
+        }
       }
     }
     
@@ -48,7 +50,7 @@ const MyProfile = () => {
   return (
     <Profile 
         name = "My"
-        desc = "Welcome to your personalized profile page"
+        desc = "Welcome to your personalized profile page. Explore your amazing prompts."
         data = {posts}
         handleEdit = {handleEdit}
         handleDelete = {handleDelete}
